matched each script

+++ Assets/!My_MemoPlatformer/MMP_Characters/AbilitiesSystem/Abilities_StateScripts/Landing.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace My_MemoPlatformer
@@ -7,10 +5,9 @@ namespace My_MemoPlatformer
     [CreateAssetMenu(fileName = "New state", menuName = "My_MemoPlatformer/AbilityData/Landing")]
     public class Landing : CharacterAbility
     {
-
         public override void OnEnter(CharacterState characterState, Animator animator, AnimatorStateInfo stateInfo)
         {
-            animator.SetBool(HashManager.Instance.arrMainParams[(int)(int)MainParameterType.Move], false);
+            animator.SetBool(HashManager.Instance.arrMainParams[(int)MainParameterType.Move], false);
         }
 
         public override void OnExit(CharacterState characterState, Animator animator, AnimatorStateInfo stateInfo)
